# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Washer & Dryer Rental CRM API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export type ClientStatus = typeof ClientStatus[keyof typeof ClientStatus];


export const ClientStatus = {
  lead: 'lead',
  active_renter: 'active_renter',
  past_customer: 'past_customer',
} as const;

export interface Client {
  id: number;
  name: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  status: ClientStatus;
  /** @nullable */
  squareCustomerId?: string | null;
  /** @nullable */
  notes?: string | null;
  activeRentalCount?: number;
  createdAt: string;
}

export type ClientInputStatus = typeof ClientInputStatus[keyof typeof ClientInputStatus];


export const ClientInputStatus = {
  lead: 'lead',
  active_renter: 'active_renter',
  past_customer: 'past_customer',
} as const;

export interface ClientInput {
  /** @minLength 1 */
  name: string;
  address?: string;
  phone?: string;
  email?: string;
  status?: ClientInputStatus;
  notes?: string;
}

export type ClientUpdateStatus = typeof ClientUpdateStatus[keyof typeof ClientUpdateStatus];


export const ClientUpdateStatus = {
  lead: 'lead',
  active_renter: 'active_renter',
  past_customer: 'past_customer',
} as const;

export interface ClientUpdate {
  name?: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  status?: ClientUpdateStatus;
  /** @nullable */
  squareCustomerId?: string | null;
  /** @nullable */
  notes?: string | null;
}

export type LeadStage = typeof LeadStage[keyof typeof LeadStage];


export const LeadStage = {
  contacted: 'contacted',
  agreement_sent: 'agreement_sent',
  term_selected: 'term_selected',
  converted: 'converted',
} as const;

export interface Lead {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  clientPhone?: string | null;
  /** @nullable */
  clientEmail?: string | null;
  stage: LeadStage;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  updatedAt?: string;
}

export type LeadInputStage = typeof LeadInputStage[keyof typeof LeadInputStage];


export const LeadInputStage = {
  contacted: 'contacted',
  agreement_sent: 'agreement_sent',
  term_selected: 'term_selected',
  converted: 'converted',
} as const;

export interface LeadInput {
  clientId: number;
  stage: LeadInputStage;
  notes?: string;
}

export type LeadUpdateStage = typeof LeadUpdateStage[keyof typeof LeadUpdateStage];


export const LeadUpdateStage = {
  contacted: 'contacted',
  agreement_sent: 'agreement_sent',
  term_selected: 'term_selected',
  converted: 'converted',
} as const;

export interface LeadUpdate {
  stage?: LeadUpdateStage;
  /** @nullable */
  notes?: string | null;
}

export interface Rental {
  id: number;
  clientId: number;
  /** @nullable */
  clientName?: string | null;
  /** @nullable */
  unitDescription?: string | null;
  startDate: string;
  termMonths: number;
  monthlyRate: number;
  monthsRemaining: number;
  isExpiringSoon: boolean;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
}

export interface RentalInput {
  clientId: number;
  unitDescription?: string;
  startDate: string;
  /** @minimum 1 */
  termMonths: number;
  /** @minimum 0 */
  monthlyRate: number;
  notes?: string;
}

export interface RentalUpdate {
  /** @nullable */
  unitDescription?: string | null;
  startDate?: string;
  /** @minimum 1 */
  termMonths?: number;
  /** @minimum 0 */
  monthlyRate?: number;
  /** @nullable */
  notes?: string | null;
}

export interface RentalAgreement {
  id: number;
  clientId: number;
  /** @nullable */
  rentalId?: number | null;
  fileName: string;
  objectPath: string;
  createdAt: string;
}

export interface AgreementInput {
  /** @nullable */
  rentalId?: number | null;
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  objectPath: string;
}

export interface SquarePayment {
  id: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  squarePaymentId?: string;
  amount: number;
  status: string;
  /** @nullable */
  description?: string | null;
  paymentDate: string;
  createdAt: string;
}

export interface SquareInvoice {
  id: number;
  /** @nullable */
  clientId?: number | null;
  /** @nullable */
  clientName?: string | null;
  squareInvoiceId?: string;
  amount: number;
  status: string;
  /** @nullable */
  dueDate?: string | null;
  createdAt: string;
}

export interface SquareSyncResult {
  customersImported: number;
  paymentsImported: number;
  invoicesImported: number;
}

export interface SquareStatus {
  connected: boolean;
  /** @nullable */
  merchantId?: string | null;
}

export interface DashboardStats {
  activeRentals: number;
  expiringSoon: number;
  openLeads: number;
  lastMonthSales: number;
  totalRevenue: number;
  totalClients: number;
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type ListClientsParams = {
search?: string;
/**
 * @nullable
 */
status?: string | null;
};

export type ListLeadsParams = {
/**
 * @nullable
 */
stage?: string | null;
/**
 * @nullable
 */
clientId?: number | null;
};

export type ListRentalsParams = {
/**
 * @nullable
 */
clientId?: number | null;
/**
 * @nullable
 */
expiringSoon?: boolean | null;
};

export type ListSquarePaymentsParams = {
/**
 * @nullable
 */
clientId?: number | null;
};

export type ListSquareInvoicesParams = {
/**
 * @nullable
 */
clientId?: number | null;
};