# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: Washer & Dryer Rental CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List clients
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: ["string", "null"]
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a client
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          description: Not found
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Update a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete a client
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Leads ─────────────────────────────────────────────────────────────────
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: stage
          in: query
          schema:
            type: ["string", "null"]
        - name: clientId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create a lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
        "404":
          description: Not found
    patch:
      operationId: updateLead
      tags: [leads]
      summary: Update a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LeadUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete a lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Rentals ───────────────────────────────────────────────────────────────
...
  /clients/{clientId}/agreements:
    get:
      operationId: listAgreements
      tags: [agreements]
      summary: List rental agreement files for a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of agreements
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RentalAgreement"
    post:
      operationId: createAgreement
      tags: [agreements]
      summary: Attach a rental agreement file to a client
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgreementInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RentalAgreement"
  /clients/{clientId}/agreements/{id}:
    delete:
      operationId: deleteAgreement
      tags: [agreements]
      summary: Delete a rental agreement attachment
      parameters:
        - name: clientId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Square ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Client:
      type: object
      required: [id, name, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [lead, active_renter, past_customer]
        squareCustomerId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        activeRentalCount:
          type: integer
        createdAt:
          type: string
    ClientInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
        address:
          type: string
        phone:
          type: string
        email:
          type: string
        status:
          type: string
          enum: [lead, active_renter, past_customer]
          default: lead
        notes:
          type: string
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
        address:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        status:
          type: string
          enum: [lead, active_renter, past_customer]
        squareCustomerId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]

    # ── Lead ─────────────────────────────────────────────────────────────────
    Lead:
      type: object
      required: [id, clientId, stage, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        clientName:
          type: ["string", "null"]
        clientPhone:
          type: ["string", "null"]
        clientEmail:
          type: ["string", "null"]
        stage:
          type: string
          enum: [contacted, agreement_sent, term_selected, converted]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    LeadInput:
      type: object
      required: [clientId, stage]
      properties:
        clientId:
          type: integer
        stage:
          type: string
          enum: [contacted, agreement_sent, term_selected, converted]
          default: contacted
        notes:
          type: string
    LeadUpdate:
      type: object
      properties:
        stage:
          type: string
          enum: [contacted, agreement_sent, term_selected, converted]
        notes:
          type: ["string", "null"]

    # ── Rental ───────────────────────────────────────────────────────────────
...
    RentalAgreement:
      type: object
      required: [id, clientId, fileName, objectPath, createdAt]
      properties:
        id:
          type: integer
        clientId:
          type: integer
        rentalId:
          type: ["integer", "null"]
        fileName:
          type: string
        objectPath:
          type: string
        createdAt:
          type: string
    AgreementInput:
      type: object
      required: [fileName, objectPath]
      properties:
        rentalId:
          type: ["integer", "null"]
        fileName:
          type: string
          minLength: 1
        objectPath:
          type: string
          minLength: 1

    # ── Square ───────────────────────────────────────────────────────────────
...