# lib/api-zod/src/generated/types/agreementInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Washer & Dryer Rental CRM API
 * OpenAPI spec version: 0.2.0
 */

export interface AgreementInput {
  /** @nullable */
  rentalId?: number | null;
  /** @minLength 1 */
  fileName: string;
  /** @minLength 1 */
  objectPath: string;
}