# lib/api-zod/src/generated/types/leadUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Washer & Dryer Rental CRM API
 * OpenAPI spec version: 0.2.0
 */
import type { LeadUpdateStage } from './leadUpdateStage';

export interface LeadUpdate {
  stage?: LeadUpdateStage;
  /** @nullable */
  notes?: string | null;
}