# lib/api-zod/src/generated/types/rentalAgreement.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Washer & Dryer Rental CRM API
 * OpenAPI spec version: 0.2.0
 */

export interface RentalAgreement {
  id: number;
  clientId: number;
  /** @nullable */
  rentalId?: number | null;
  fileName: string;
  objectPath: string;
  createdAt: string;
}